Committing unit tests and functionality
WindowsWrapper project handles the interop side of interacting with Windows binaries so the other projects don't need to worry about the details of method signatures.

UnitTests project tests each part of the projects to make sure they function as expected.

PixelClicker is the main entry point for the application and pieces the other projects together into a single application.

MouseWrapper